Fix: Enable file picker with storage permissions and build compatibility
- Fixed win32 package version (5.1.1) for Dart 3.3.4 compatibility
- Disabled Jetifier to resolve build errors
- Added storage permissions to AndroidManifest.xml
- File picker now supports ALL file types (FileType.any)
- Created install_and_run.bat script for easy deployment
- APK built successfully: 53.2MB
- Ready for testing on emulator

+++ lib/features/backup/services/backup_service.dart
@@ -190,7 +190,7 @@ class BackupService {
         
         try {
           final encryptedData = await _storage.readEncrypted(meta.id);
-          totalSize += encryptedData.cipher.length as int;
+          totalSize += encryptedData.cipher.length;
         } catch (e) {
           // Skip files that can't be read
         }
